+++ lib-gphotos/internal/uploader/errors.go
@@ -1,0 +1,7 @@
+package uploader
+
+import "golang.org/x/xerrors"
+
+var (
+	ErrNilStore = xerrors.New("store can't be nil if Resume is enable")
+)